Refactor shared utilities into a common subpackage
