+++ .github/workflows/helmcluster.yaml
@@ -1,16 +1,14 @@
-name: "KubeCluster"
+name: "HelmCluster"
 on:
   pull_request:
     paths:
       - "ci/**"
-      - "dask_kubernetes/classic/**"
-      - "dask_kubernetes/tests/**"
+      - "dask_kubernetes/helm/**"
       - "dask_kubernetes/*"
   push:
     paths:
       - "ci/**"
-      - "dask_kubernetes/classic/**"
-      - "dask_kubernetes/tests/**"
+      - "dask_kubernetes/helm/**"
       - "dask_kubernetes/*"
 
 
@@ -32,4 +30,4 @@ jobs:
       - name: Run tests
         env:
           KUBECONFIG: .pytest-kind/pytest-kind/kubeconfig
-        run: pytest dask_kubernetes/tests
+        run: pytest dask_kubernetes/common/tests dask_kubernetes/helm/tests
