+++ dask_kubernetes/operator/operator.py
@@ -7,8 +7,8 @@ import kubernetes_asyncio as kubernetes
 
 from uuid import uuid4
 
-from dask_kubernetes.auth import ClusterAuth
-from dask_kubernetes.networking import (
+from dask_kubernetes.common.auth import ClusterAuth
+from dask_kubernetes.common.networking import (
     get_scheduler_address,
 )
